requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
python-docx>=1.1.2
openpyxl>=3.1
xlsxwriter>=3.2
